+++ lib/models/bridge_signal.dart
@@ -9,56 +9,108 @@ abstract class BaseBridgeSignal<V> extends Signal<V?> {
   void connect(InheritedTrinityScope scope);
 }
 
-///[N] is the refrence node where we'll search the reference signal.
+/// A bridge that mirrors a parent node's signal with no transformation.
 ///
-///[S] is the type of the reference signal.
+/// [N] is the parent node type where the source signal lives.
+/// [S] is both the source signal type and the bridge's value type (V == S).
 ///
-///[V] is the type of the bridge signal.
+/// The value setter writes directly back to the parent signal.
 ///
-///When you create a bridge on your child Node you must register inside the bridges using
-///[registerBridge] or [registerManyBridges].
-///
-///You should never expose it to widgets, in order to read the state you need to use
-///[readableSignal] through any [SignalBuilder] available.
-class BridgeSignal<N extends NodeInterface, S, V> extends BaseBridgeSignal<V> {
-  StreamSubscription<S>? _subscription;
+/// ```dart
+/// late final edad = registerSignal(
+///   BridgeSignal(select: (FormNode node) => node.edad),
+/// );
+/// ```
+class BridgeSignal<N extends NodeInterface, V> extends BaseBridgeSignal<V> {
+  StreamSubscription<V>? _subscription;
   late final N _parentNode;
 
-  ///[select] is a function that will select the reference signal from the node.
-  final ReadableSignal<S> Function(N node) select;
+  final Signal<V> Function(N node) _select;
 
-  ///[transform] is a function that will transform the value of the reference signal
-  ///and then emit it.
-  final V? Function(S value) transform;
-
-  ///[_update] is a function that will update the value of the bridge signal.
-  ///
-  ///You should use Signal function value(newValue) to update
-  ///
-  ///You're not supposed to call this function outside the BridgeSignal constructor.
-  ///This is only a reference to update value directly from father.
-  final Function(N node, V? value) _update;
-
-  BridgeSignal({
-    required this.select,
-    required this.transform,
-    required Function(N, V?) update,
-  }) : _update = update,
-       super(null);
+  BridgeSignal({required Signal<V> Function(N) select})
+    : _select = select,
+      super(null);
 
   @override
-  set value(V? newValue) => _update(_parentNode, newValue);
+  set value(covariant V newValue) {
+    _select(_parentNode).value = newValue;
+  }
+
+  @override
+  V get value => _select(_parentNode).value;
+
   @override
   @protected
   void connect(InheritedTrinityScope scope) {
     final node = _parentNode = scope.findByType<N>();
-    final parentSignal = select(node);
-    final initialValue = transform(parentSignal.value);
+    final parentSignal = _select(node);
+
+    emit(parentSignal.value);
+
+    _subscription = parentSignal.streamTriggerImmediatly.listen((data) {
+      emit(data);
+    });
+  }
+
+  @override
+  void dispose() {
+    _subscription?.cancel();
+    super.dispose();
+  }
+}
+
+/// A bridge that transforms a parent signal's value into a different type.
+///
+/// [N] is the parent node type.
+/// [S] is the source signal type.
+/// [V] is the transformed value type (can differ from S, value is nullable).
+///
+/// ```dart
+/// late final order = registerSignal(
+///   TransformBridgeSignal(
+///     select: (OrdersNode node) => node.orders,
+///     transform: (orders) => orders.firstWhereOrNull((o) => o.id == id),
+///     update: (node, value) {
+///       if (value == null) return;
+///       node.updateOrder(value);
+///     },
+///   ),
+/// );
+/// ```
+class TransformBridgeSignal<N extends NodeInterface, S, V>
+    extends BaseBridgeSignal<V> {
+  StreamSubscription<S>? _subscription;
+  late final N _parentNode;
+
+  final Signal<S> Function(N node) _select;
+  final V? Function(S value)? _transform;
+  final void Function(N node, V? value) _update;
+
+  TransformBridgeSignal({
+    required Signal<S> Function(N) select,
+    required V? Function(S)? transform,
+    required void Function(N, V?) update,
+  }) : _select = select,
+       _transform = transform,
+       _update = update,
+       super(null);
+
+  @override
+  set value(V? newValue) {
+    _update(_parentNode, newValue);
+  }
+
+  @override
+  @protected
+  void connect(InheritedTrinityScope scope) {
+    final node = _parentNode = scope.findByType<N>();
+    final parentSignal = _select(node);
+    final initialValue = _transform?.call(parentSignal.value);
 
     emit(initialValue);
 
     _subscription = parentSignal.streamTriggerImmediatly.listen((data) {
-      final transformed = transform(data);
+      final transformed = _transform?.call(data);
       emit(transformed);
     });
   }