+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-## 0.0.1
+## 0.1.0
 
-* TODO: Describe initial release.
+* Separate BridgeSignal into [BridgeSignal] and [TransformBridgeSignal]
+* Shortened [Signal] function in [SignalBuilder] to allow returning the signal, not the readable
+* Fixed generator to generate [ReadableNode] for nodes with signals
+
+<!-- ## 0.2.0
+
+* TODO: Describe changes in 0.2.0 -->